Todays Progress - (we have a bug to fix in custom_model)
The model is showing training logs but didn't learn anything significant. I checked the models parameters and it has none. Probably because it need to hold the trainable models in self and not a containers classes that holds them.

Tomorrows job - Fix this by making the containers inherit from nn.module OR hold them directly in the custom model self.